requirements.txt
Django==4.2.18
gunicorn==19.9.0
pip==23.3.2
djangorestframework==3.15.2
urllib3==2.3.0
sqlparse==0.5.3
PyJWT==2.10.1
